FIX: html mejor organizado, cambié FloatField en forms.py por IntegerField

+++ prueba_tecnica/home/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 
 class CabidaForm(forms.Form):
-    x = forms.FloatField(label='Lado X', min_value=0)
-    y = forms.FloatField(label='Lado Y', min_value=0)
-    a = forms.FloatField(label='Lado A', min_value=0)
-    b = forms.FloatField(label='Lado B', min_value=0)
+    x = forms.IntegerField(label='Lado X', min_value=0)
+    y = forms.IntegerField(label='Lado Y', min_value=0)
+    a = forms.IntegerField(label='Lado A', min_value=0)
+    b = forms.IntegerField(label='Lado B', min_value=0)
